README.CD
This is a C++ client-server application that uses the Command-pattern for interaction. To allow communication, the local port needs to be allowed through the firewall on both the server and client sides. The server creates a listening socket and handles incoming client connections. Users can send messages to other clients via the server. The project requires CMake to build.